Android/Downloader update for 2.0
